Split arrow part and parquet part
